paperwurksapi/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.3.1
    # via kombu
annotated-types==0.7.0
    # via pydantic
asgiref==3.9.2
    # via
    #   django
    #   django-cors-headers
asttokens==3.0.0
    # via stack-data
billiard==4.2.2
    # via celery
boto3==1.35.99
    # via -r requirements.in
botocore==1.35.99
    # via
    #   boto3
    #   s3transfer
build==1.3.0
    # via pip-tools
celery==5.5.3
    # via -r requirements.in
certifi==2025.10.5
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   mkdocs
    #   pip-tools
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
coverage==7.6.12
    # via -r requirements.in
cryptography==46.0.2
    # via -r requirements.in
decorator==5.2.1
    # via
    #   ipdb
    #   ipython
dj-database-url==2.3.0
    # via -r requirements.in
django==5.1.13
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-configurations
    #   django-cors-headers
    #   django-filter
    #   django-model-utils
    #   django-ninja
    #   django-storages
django-configurations==2.5.1
    # via -r requirements.in
django-cors-headers==4.9.0
    # via -r requirements.in
django-filter==24.3
    # via -r requirements.in
django-model-utils==5.0.0
    # via -r requirements.in
django-ninja==1.4.3
    # via -r requirements.in
django-storages==1.14.6
    # via -r requirements.in
django-unique-upload==0.2.1
    # via -r requirements.in
executing==2.2.1
    # via stack-data
factory-boy==3.3.3
    # via -r requirements.in
faker==37.8.0
    # via factory-boy
flake8==7.1.2
    # via -r requirements.in
ghp-import==2.1.0
    # via mkdocs
gunicorn==23.0.0
    # via -r requirements.in
idna==3.10
    # via requests
iniconfig==2.1.0
    # via pytest
ipdb==0.13.13
    # via -r requirements.in
ipython==8.30.0
    # via
    #   -r requirements.in
    #   ipdb
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via mkdocs
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.5.4
    # via celery
markdown==3.9
    # via mkdocs
markupsafe==3.0.3
    # via
    #   jinja2
    #   mkdocs
matplotlib-inline==0.1.7
    # via ipython
mccabe==0.7.0
    # via flake8
mergedeep==1.3.4
    # via
    #   mkdocs
    #   mkdocs-get-deps
mkdocs==1.6.1
    # via -r requirements.in
mkdocs-get-deps==0.2.0
    # via mkdocs
mock==5.1.0
    # via -r requirements.in
newrelic==10.3.1
    # via -r requirements.in
packaging==25.0
    # via
    #   build
    #   gunicorn
    #   kombu
    #   mkdocs
    #   pytest
parso==0.8.5
    # via jedi
pathspec==0.12.1
    # via mkdocs
pexpect==4.9.0
    # via ipython
pip-tools==7.5.1
    # via -r requirements.in
platformdirs==4.4.0
    # via mkdocs-get-deps
pluggy==1.6.0
    # via pytest
prompt-toolkit==3.0.52
    # via
    #   click-repl
    #   ipython
psycopg2-binary==2.9.10
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pycodestyle==2.12.1
    # via flake8
pycparser==2.23
    # via cffi
pydantic==2.11.9
    # via django-ninja
pydantic-core==2.33.2
    # via pydantic
pyflakes==3.2.0
    # via flake8
pygments==2.19.2
    # via
    #   ipython
    #   pytest
pyjwt==2.10.1
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.4.2
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.9.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
    #   ghp-import
python-decouple==3.8
    # via -r requirements.in
pyyaml==6.0.3
    # via
    #   mkdocs
    #   mkdocs-get-deps
    #   pyyaml-env-tag
pyyaml-env-tag==1.1
    # via mkdocs
redis==6.4.0
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
s3transfer==0.10.4
    # via boto3
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
stack-data==0.6.3
    # via ipython
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.15.0
    # via
    #   dj-database-url
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via
    #   faker
    #   kombu
urllib3==2.5.0
    # via
    #   botocore
    #   requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
watchdog==6.0.0
    # via mkdocs
wcwidth==0.2.14
    # via prompt-toolkit
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
